+++ build.gradle
@@ -10,7 +10,7 @@ plugins {
 group 'org.example'
 version '1.5.0.0'
 
-sourceCompatibility = 1.8
+sourceCompatibility = 11
 
 repositories {
     mavenCentral()
@@ -34,7 +34,7 @@ dependencies {
     implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.8.10'
     implementation group: 'org.json', name: 'json', version: '20230618'
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.9.2'
-    testImplementation group: 'io.mockk', name: 'mockk', version: '1.13.4'
+    testImplementation group: 'io.mockk', name: 'mockk', version: '1.13.8'
     testImplementation 'com.willowtreeapps.assertk:assertk-jvm:0.25'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.9.2'
     testImplementation 'info.solidsoft.gradle.pitest:gradle-pitest-plugin:1.9.11'
@@ -49,10 +49,10 @@ intellij {
 }
 
 compileKotlin {
-    kotlinOptions.jvmTarget = "1.8"
+    kotlinOptions.jvmTarget = "11"
 }
 compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.8"
+    kotlinOptions.jvmTarget = "11"
 }
 patchPluginXml {
     sinceBuild = '193'
