+++ src/test/kotlin/inigo/KotlinObjectMotherTemplateTest.kt
@@ -87,43 +87,6 @@ class KotlinObjectMotherTemplateTest {
         assertThatWorksWithType("boolean", "faker.bool().bool()")
     }
 
-    @Test
-    fun `strings are generated from list of strings`() {
-        assertTrue { listOf("faker.ancient().god()",
-                "faker.ancient().primordial()",
-                "faker.ancient().titan()",
-                "faker.artist().name()",
-                "faker.backToTheFuture().character()",
-                "faker.backToTheFuture().quote()",
-                "faker.beer().name()",
-                "faker.buffy().characters()",
-                "faker.buffy().quotes()",
-                "faker.chuckNorris().fact()",
-                "faker.dragonBall().character()",
-                "faker.funnyName().name()",
-                "faker.friends().character()",
-                "faker.friends().quote()",
-                "faker.gameOfThrones().character()",
-                "faker.gameOfThrones().quote()",
-                "faker.hipster().word()",
-                "faker.hitchhikersGuideToTheGalaxy().character()",
-                "faker.hitchhikersGuideToTheGalaxy().marvinQuote()",
-                "faker.hitchhikersGuideToTheGalaxy().quote()",
-                "faker.lebowski().quote()",
-                "faker.howIMetYourMother().character()",
-                "faker.howIMetYourMother().catchPhrase()",
-                "faker.howIMetYourMother().highFive()",
-                "faker.howIMetYourMother().quote()",
-                "faker.lordOfTheRings().location()",
-                "faker.princessBride().quote()",
-                "faker.princessBride().character()",
-                "faker.rickAndMorty().quote()",
-                "faker.rickAndMorty().character()",
-                "faker.slackEmoji().activity()",
-                "faker.superhero().name()",
-                "faker.yoda().quote()").contains(KotlinFakeValuesGenerator().strings[0]) }
-    }
-
     private fun assertThatWorksWithType(type: String, expectedGenerator: String) {
         val sut = KotlinObjectMotherTemplate(KotlinFakeValuesGenerator())
