+++ src/test/kotlin/inigo/objectMotherCreator/features/JavaFeaturesTests.kt
@@ -4,19 +4,17 @@ import inigo.objectMotherCreator.application.JavaFileCreator
 import inigo.objectMotherCreator.application.ObjectMotherCreator
 import inigo.objectMotherCreator.application.infoholders.ClassInfo
 import inigo.objectMotherCreator.application.template.JavaObjectMotherTemplate
-import inigo.objectMotherCreator.application.template.KotlinObjectMotherTemplate
+import inigo.objectMotherCreator.application.values.ClassMapping
+import inigo.objectMotherCreator.application.values.DefaultMappings
 import inigo.objectMotherCreator.application.values.FakeValuesGenerator
 import inigo.objectMotherCreator.application.values.JavaFakeValuesGenerator
-import inigo.objectMotherCreator.application.values.KotlinFakeValuesGenerator
 import inigo.objectMotherCreator.givenStandartStateOptions
 import inigo.objectMotherCreator.infraestructure.IdeaShits
 import inigo.objectMotherCreator.model.infoExtractor.om.*
-import io.mockk.MockKAnnotations
-import io.mockk.every
+import io.mockk.*
 import io.mockk.impl.annotations.MockK
 import io.mockk.impl.annotations.SpyK
-import io.mockk.mockk
-import io.mockk.verify
+import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.sql.Timestamp
@@ -40,11 +38,17 @@ class JavaFeaturesTests {
     lateinit var fileCreator: JavaFileCreator
 
     @SpyK
-    var fakeValuesGenerator: FakeValuesGenerator = JavaFakeValuesGenerator()
+    var defaults: DefaultMappings = spyk(DefaultMappings())
+
+    @MockK
+    lateinit var fakeValuesGenerator: FakeValuesGenerator
 
     @BeforeEach
     fun setup() {
         MockKAnnotations.init(this)
+        mockkObject(FakeValuesGenerator.Companion)
+        defaults.mappings = setOf(ClassMapping(generator = listOf ("faker.howIMetYourMother().highFive()"), className = "String")) + defaults.mappings
+        fakeValuesGenerator = JavaFakeValuesGenerator(defaults)
         every { fileCreator.buildFile(any(), any(), any(), any()) } returns Unit
         every { ideShits.findClass("java.lang.String") } returns null
         every { ideShits.findClass("java.util.List") } returns null
@@ -53,7 +57,11 @@ class JavaFeaturesTests {
         every { ideShits.findClass("java.time.Instant") } returns null
         every { ideShits.findClass("java.sql.Timestamp") } returns null
         every { fileCreator.createdFileName() } returns "createdObjectMother"
-        every { fakeValuesGenerator.randomString() } returns "faker.howIMetYourMother().highFive()"
+    }
+
+    @AfterEach
+    fun shutdown() {
+        unmockkAll()
     }
 
     @Test
