+++ src/test/kotlin/inigo/objectMotherCreator/features/KotlinFeaturesTests.kt
@@ -4,14 +4,14 @@ import inigo.objectMotherCreator.application.JavaFileCreator
 import inigo.objectMotherCreator.application.ObjectMotherCreator
 import inigo.objectMotherCreator.application.infoholders.ClassInfo
 import inigo.objectMotherCreator.application.template.KotlinObjectMotherTemplate
-import inigo.objectMotherCreator.application.values.FakeValuesGenerator
-import inigo.objectMotherCreator.application.values.KotlinFakeValuesGenerator
+import inigo.objectMotherCreator.application.values.*
 import inigo.objectMotherCreator.givenStandartStateOptions
 import inigo.objectMotherCreator.infraestructure.IdeaShits
 import inigo.objectMotherCreator.model.infoExtractor.om.*
 import io.mockk.*
 import io.mockk.impl.annotations.MockK
 import io.mockk.impl.annotations.SpyK
+import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.sql.Timestamp
@@ -35,11 +35,17 @@ class KotlinFeaturesTests {
     lateinit var fileCreator: JavaFileCreator
 
     @SpyK
-    var fakeValuesGenerator: FakeValuesGenerator = KotlinFakeValuesGenerator()
+    var defaults: DefaultMappings = spyk(DefaultMappings())
+
+    @MockK
+    lateinit var fakeValuesGenerator: FakeValuesGenerator
 
     @BeforeEach
     fun setup() {
         MockKAnnotations.init(this)
+        mockkObject(FakeValuesGenerator.Companion)
+        defaults.mappings = setOf(ClassMapping(generator = listOf ("faker.howIMetYourMother().highFive()"), className = "String")) + defaults.mappings
+        fakeValuesGenerator = KotlinFakeValuesGenerator(defaults)
         every { fileCreator.buildFile(any(), any(), any(), any()) } returns Unit
         every { ideShits.findClass("java.lang.String") } returns null
         every { ideShits.findClass("java.util.List") } returns null
@@ -48,7 +54,11 @@ class KotlinFeaturesTests {
         every { ideShits.findClass("java.time.Instant") } returns null
         every { ideShits.findClass("java.sql.Timestamp") } returns null
         every { fileCreator.createdFileName() } returns "createdObjectMother"
-        every { fakeValuesGenerator.randomString() } returns "faker.howIMetYourMother().highFive()"
+    }
+
+    @AfterEach
+    fun shutdown() {
+        unmockkAll()
     }
 
     @Test
